requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
pandas
APScheduler
httpx
loguru
pydantic_settings
passlib[bcrypt]
pyjwt
pydantic[email]
python-multipart